perf(ui): remove first-render startup waits
New Session and Chat start route loading without separate mandatory-helper
downloads. Call the canonical session URL builder directly and eagerly load
the initial-location resolver, preserving authentication, Gateway defaults,
first-run decisions, pending navigation, and abort/disposal ownership.

Remove two JavaScript assets and 26 net production lines without adding a
cache, configuration, dependency, or compatibility path. Pre-refresh Node 24
proof passed 187 tests across 16 files and all 22 changed checks. The retained
96-navigation HTTP/1.1 and HTTP/2 comparison favored the candidate in every
paired cold navigation; initial JavaScript grew 470 gzip bytes within the
unchanged budget, with byte-identical initial CSS.
