+++ ui/src/app-session-path-builder.ts
@@ -1,4 +1,7 @@
-import type { ControlUiSessionNamespace } from "@openclaw/session-url-contract";
+import {
+  buildControlUiSessionPath,
+  type ControlUiSessionNamespace,
+} from "@openclaw/session-url-contract";
 
 type SessionPathDetails = {
   displayName?: string | null;
@@ -7,14 +10,6 @@ type SessionPathDetails = {
   shortIdLength?: number;
 };
 
-type SessionPathBuilder = typeof import("@openclaw/session-url-contract").buildControlUiSessionPath;
-
-let builder: SessionPathBuilder | undefined;
-
-export function setSessionPathBuilder(next: SessionPathBuilder): void {
-  builder = next;
-}
-
 export function pathForSession(
   face: ControlUiSessionNamespace,
   agentId: string,
@@ -22,16 +17,14 @@ export function pathForSession(
   basePath = "",
   details: SessionPathDetails = {},
 ): string | null {
-  return (
-    builder?.({
-      namespace: face,
-      sessionKey,
-      fallbackAgentId: agentId,
-      basePath,
-      displayName: details.displayName ?? undefined,
-      exactKey: details.exactKey,
-      mainKey: details.mainKey ?? undefined,
-      shortIdLength: details.shortIdLength,
-    }) ?? null
-  );
+  return buildControlUiSessionPath({
+    namespace: face,
+    sessionKey,
+    fallbackAgentId: agentId,
+    basePath,
+    displayName: details.displayName ?? undefined,
+    exactKey: details.exactKey,
+    mainKey: details.mainKey ?? undefined,
+    shortIdLength: details.shortIdLength,
+  });
 }
