+++ ui/src/app/bootstrap.ts
@@ -14,7 +14,6 @@ import {
   type ApplicationRouter,
   type RouteId,
 } from "../app-routes.ts";
-import { setSessionPathBuilder } from "../app-session-path-builder.ts";
 import {
   SIDEBAR_SESSION_NAV_COLLAPSE_QUERY,
   sessionRefFromPath,
@@ -34,6 +33,7 @@ import {
 } from "../pages/model-setup/first-run.ts";
 import { createAgentSelectionCapability } from "./agent-selection.ts";
 import { resolveControlUiDocumentMode, type ControlUiDocumentMode } from "./approval-deep-link.ts";
+import { resolveInitialApplicationLocation } from "./bootstrap-location.ts";
 import { createApplicationTheme } from "./bootstrap-theme.ts";
 import { createBrowserHistory, resolveControlUiPaths } from "./browser.ts";
 import { createChatAttachmentHandoff } from "./chat-attachment-handoff.ts";
@@ -132,19 +132,13 @@ export type ApplicationRuntime = {
   stop: () => void;
 };
 
-type BootstrapApplicationDependencies = {
-  sessionPathBuilderReady?: Promise<void>;
-};
-
 type PendingRouterStartNavigation = {
   routeId: RouteId;
   location: RouteLocation;
   mode: "push" | "replace";
 };
 
-export function bootstrapApplication(
-  dependencies: BootstrapApplicationDependencies = {},
-): ApplicationRuntime {
+export function bootstrapApplication(): ApplicationRuntime {
   const history = createBrowserHistory();
   const startupLocation = history.location();
   const [basePath, resourceBasePath] = resolveControlUiPaths(
@@ -203,15 +197,6 @@ export function bootstrapApplication(
   const startsApplicationRouter = documentMode === null && focusLocation === null;
   const firstRunDefaultLanding =
     startsApplicationRouter && isDefaultChatLanding(applicationLocation, basePath, routeIdFromPath);
-  const sessionPathBuilderReady =
-    dependencies.sessionPathBuilderReady ??
-    (documentMode ||
-    (focusLocation?.status === "valid" && focusLocation.target.kind !== "dashboard")
-      ? Promise.resolve()
-      : import("@openclaw/session-url-contract").then((contract) => {
-          setSessionPathBuilder(contract.buildControlUiSessionPath);
-        }));
-
   const hasPendingGateway = startup.pendingGatewayUrl !== null;
   const gateway = createApplicationGateway(
     settings,
@@ -241,21 +226,18 @@ export function bootstrapApplication(
   const initialLocationReady = (
     documentMode || focusLocation
       ? Promise.resolve(applicationLocation)
-      : Promise.all([sessionPathBuilderReady, import("./bootstrap-location.ts")]).then(
-          ([, location]) =>
-            location.resolveInitialApplicationLocation({
-              location: applicationLocation,
-              basePath,
-              sessionKey: settings.sessionKey,
-              gateway,
-              agentsList: () => agents.state.agentsList,
-              selectedAgentId: settings.selectedAgentId,
-              signal: startupLifecycle.signal,
-            }),
-        )
+      : resolveInitialApplicationLocation({
+          location: applicationLocation,
+          basePath,
+          sessionKey: settings.sessionKey,
+          gateway,
+          agentsList: () => agents.state.agentsList,
+          selectedAgentId: settings.selectedAgentId,
+          signal: startupLifecycle.signal,
+        })
   ).catch((error: unknown) => {
-    // stop() aborts an eager unscoped-session lookup even when start() returns
-    // at the lazy-chunk guard, so consume that teardown-only rejection here.
+    // stop() aborts eager session lookups even before start() reaches location
+    // resolution, so consume that teardown-only rejection here.
     if (startupLifecycle.signal.aborted) {
       return applicationLocation;
     }
@@ -496,7 +478,6 @@ export function bootstrapApplication(
           return () => gateway.stop();
         },
         () => startGatewayPageActivation(gateway, document, window),
-        () => sessionPathBuilderReady,
       ];
       // Resolve first-run setup before routing: the default Chat route owns the
       // workspace graph, which setup users would otherwise fetch and discard.
